requirements.txt
streamlit~=1.23.1
requests~=2.31.0
pandas~=2.0.2
numpy~=1.25.0
joblib~=1.2.0
scikit-learn~=1.2.2
xgboost~=1.7.6
Pillow~=9.5.0